pyproject.toml
[project]
name = "lacommunaute"
version = "2.24.0"
description = "La communauté de l'inclusion"
authors = []
license = {text = "ETALAB 2.0"}
requires-python = ">=3.13"
readme = "README.md"
dependencies = [
    "Django>=5.1",
    "python-dotenv>=1.0",
    "psycopg>=3.2",
    "django-machina>=1.2.0",
    "boto3>=1.36",
    "django-storages>=1.14",
    "httpx>=0.28",
    "django-compressor>=4.5",
    "django-libsass>=0.9",
    "sentry-sdk>=2.20",
    "django-social-share>=2.3.0",
    "django-htmx>=1.21",
    "django-taggit>=6.1",
    "django-csp>=3.8",
    "django-permissions-policy>=4.24",
    "langdetect>=1.0.9",
    "pyjwt>=2.10",
    "python-crontab>=3.2.0",
    "jwcrypto>=1.5.6",
]

[dependency-groups]
dev = [
    "pylint>=3.3",
    "pylint-django>=2.6",
    "pre-commit>=4.0",
    "coverage>=7.6",
    "django-extensions>=3.2",
    "django-debug-toolbar>=5.0",
    "Werkzeug>=3.1.3",
    "pytest-django>=4.9",
    "debugpy>=1.8",
    "factory-boy>=3.3",
    "respx>=0.22",
    "pytest-xdist>=3.6",
    "pillow>=11.1.0",
    "moto>=5.0",
    "ruff>=0.9",
    "djlint>=1.36",
    "syrupy>=4.8",
    "beautifulsoup4>=4.12",
    "html5lib~=1.1",
    "freezegun>=1.5.1",
    "pytest-mock>=3.15",
]

[tool.ruff]
line-length = 119

[tool.ruff.lint]
ignore = []
# see prefixes in https://beta.ruff.rs/docs/rules/
select = [
    "F",  # pyflakes
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "I",  # isort
]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["lacommunaute"]
lines-after-imports = 2

[tool.djlint]
profile="django"
ignore="T002,T003,T027,H006,H023,D018"
max_attribute_length=200
format_css = true
format_js = false

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["tests*.py", "test_*.py"]
addopts = [
    "--reuse-db",
    "--strict-markers",
]

[tool.pytest.ini_options.markers]
no_django_db = "mark tests that should not be marked with django_db."
